+++ src/components/layout/Content.vue
@@ -1,34 +1,13 @@
 <template>
     <div :class="$style.main">
-        <el-time-picker
-            v-model="value1"
-            :picker-options="{
-                selectableRange: '18:30:00 - 20:30:00'
-            }"
-            placeholder="任意时间点"
-        >
-        </el-time-picker>
-        <el-time-picker
-            arrow-control
-            v-model="value2"
-            :picker-options="{
-                selectableRange: '18:30:00 - 20:30:00'
-            }"
-            placeholder="任意时间点"
-        >
-        </el-time-picker>
+       <slot>
+           <!-- 我是类似于 this.props.children 的东西 -->
+        </slot>  
     </div>
 </template>
 
 <script>
-export default {
-    data() {
-        return {
-            value1: new Date(2016, 9, 10, 18, 40),
-            value2: new Date(2016, 9, 10, 18, 40)
-        };
-    }
-};
+
 </script>
 
 <style lang="scss" module>